backend/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "cryptography>=45.0.5",
    "fastapi[standard]>=0.116.1",
    "gunicorn>=23.0.0",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.1",
    "sqlmodel>=0.0.24",
    "uvicorn[standard]>=0.35.0",
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.3",
]
